+++ provider/resource_workflow_alert.go
@@ -53,6 +53,14 @@ func resourceWorkflowAlert() *schema.Resource {
 				Description: "Workflow command.",
 			},
 
+			"command_feedback_enabled": &schema.Schema{
+				Type:        schema.TypeBool,
+				Computed:    true,
+				Required:    false,
+				Optional:    true,
+				Description: "This will notify you back when the workflow is starting.",
+			},
+
 			"wait": &schema.Schema{
 				Type:        schema.TypeString,
 				Computed:    true,
@@ -318,6 +326,9 @@ func resourceWorkflowAlertCreate(ctx context.Context, d *schema.ResourceData, me
 	if value, ok := d.GetOkExists("command"); ok {
 		s.Command = value.(string)
 	}
+	if value, ok := d.GetOkExists("command_feedback_enabled"); ok {
+		s.CommandFeedbackEnabled = tools.Bool(value.(bool))
+	}
 	if value, ok := d.GetOkExists("wait"); ok {
 		s.Wait = value.(string)
 	}
@@ -390,6 +401,7 @@ func resourceWorkflowAlertRead(ctx context.Context, d *schema.ResourceData, meta
 	d.Set("slug", item.Slug)
 	d.Set("description", item.Description)
 	d.Set("command", item.Command)
+	d.Set("command_feedback_enabled", item.CommandFeedbackEnabled)
 	d.Set("wait", item.Wait)
 	d.Set("repeat_every_duration", item.RepeatEveryDuration)
 	d.Set("repeat_on", item.RepeatOn)
@@ -428,6 +440,9 @@ func resourceWorkflowAlertUpdate(ctx context.Context, d *schema.ResourceData, me
 	}
 	if d.HasChange("command") {
 		s.Command = d.Get("command").(string)
+	}
+	if d.HasChange("command_feedback_enabled") {
+		s.CommandFeedbackEnabled = tools.Bool(d.Get("command_feedback_enabled").(bool))
 	}
 	if d.HasChange("wait") {
 		s.Wait = d.Get("wait").(string)