+++ provider/resource_workflow_task_update_status.go
@@ -68,10 +68,11 @@ func resourceWorkflowTaskUpdateStatus() *schema.Resource {
 							}, false),
 						},
 						"status": &schema.Schema{
-							Description: "Value must be one of `started`, `mitigated`, `resolved`, `cancelled`.",
+							Description: "Value must be one of `in_triage`, `started`, `mitigated`, `resolved`, `cancelled`.",
 							Type:        schema.TypeString,
 							Required:    true,
 							ValidateFunc: validation.StringInSlice([]string{
+								"in_triage",
 								"started",
 								"mitigated",
 								"resolved",